Group `/set*` commands.
