+++ Aoi/twitter.py
@@ -6,7 +6,6 @@ from discord.ext import commands
 import discord
 import os
 import tweepy
-import asyncio
 import unicodedata
 import time
 
@@ -15,12 +14,11 @@ from . import (
     help_command,
     get_twitter_consumer_key,
     get_twitter_consumer_secret,
+    has_permission,
 )
 from .database import (
     get_tt_tat_tats,
-    update_twitter_template,
     update_tat_tats,
-    has_permission,
 )
 
 DATABASE_URL = get_database_url()
@@ -135,53 +133,6 @@ class Twitter(commands.Cog):
         URL = f"https://twitter.com/i/web/status/{res.data['id']}"
 
         await interaction.followup.send(f'ツイートしました。\n{URL}')
-
-    @app_commands.command()
-    @help_command()
-    @has_permission(administrator=True)
-    async def settwitter(self, interaction: discord.Interaction, clear: bool = False, help: bool = False):
-        """Change template for tweet.
-
-        Previlage of administrator is required.
-
-        > /settwitter
-        > <Template of tweet passed to jinja2.Template>
-
-        Parameters
-        ----------
-        interaction : discord.Interaction
-            _description_
-        clear : str, by default False
-            Wether to clear template.
-        help : bool, optional
-            _description_, by default False
-        """
-        GUILD_ID = interaction.guild_id
-
-        if clear:
-            template = None
-
-            update_twitter_template(DATABASE_URL, GUILD_ID, template)
-            await interaction.response.send_message("twitter_template is changed "
-                                                    f"to {template}.")
-        else:
-            await interaction.response.send_message("Input template as jinja2.Template.")
-
-            def check(m):
-                """Check if it's the same user and channel."""
-                return m.author == interaction.user and m.channel == interaction.channel
-
-            try:
-                response = await self.bot.wait_for('message', check=check, timeout=30.0)
-            except asyncio.TimeoutError:
-                await interaction.followup.send("settwitter is canceled with timeout.")
-                return
-            template = response.content
-
-            update_twitter_template(DATABASE_URL, GUILD_ID, template)
-            await interaction.followup.send("twitter_template is changed "
-                                            f"to following:\n{template}.")
-        return
 
     @app_commands.command()
     @help_command()
